Restructure graph_of_convex_sets into a subpackage
Splits the single graph_of_convex_sets.py / graph_of_convex_sets_drake.py
files into a subpackage of world_description with one file per concern:
- base.py: the abstract GraphOfConvexSets, obstacle-collection helpers,
  and the representation-agnostic free functions (translate_event_to,
  translate_free_space_to_where_condition,
  create_reference_frame_with_only_yaw_from_body).
- boxes.py: GraphOfBoundingBoxes and navigation_map_at_target.
- polygons.py: the Drake/IRIS implementation, renamed from
  DrakeGraphOfConvexSets to GraphOfConvexPolygons.
- exceptions.py: UnboundedSearchSpaceError.

__init__.py is intentionally empty; every consumer imports from the
specific submodule it needs. Updated all in-repo callers (tests, the
sage_10k experiment, graph_of_convex_sets_experiments.py, and the
example notebook) accordingly, and renamed test_gcs_drake.py to
test_gcs_polygons.py to match.

Also makes CI install drake for the semantic_digital_twin test job,
since it was made an optional dependency (not in requirements.txt) but
is required to actually exercise GraphOfConvexPolygons.

Made with the help of Claude.
