+++ .github/workflows/ci_reusable.yml
@@ -43,6 +43,15 @@ jobs:
           uv tree --depth 1
           python -c "import sys; print(sys.executable)"
 
+      - name: Install drake
+        if: ${{ inputs.lib == 'semantic_digital_twin' }}
+        env:
+          UV_CACHE_DIR: /opt/ros/.uv_cache
+        run: |
+          cd cognitive_robot_abstract_machine
+          source /opt/ros/cram-env/bin/activate
+          uv pip install drake --active
+
       - name: Restore cache
         uses: actions/cache/restore@v4
         with:
