+++ test/semantic_digital_twin_test/test_worlds/test_gcs_spawn_region.py
@@ -4,7 +4,7 @@ from semantic_digital_twin.world_description.connections import FixedConnection
 from semantic_digital_twin.world_description.world_entity import Region, Body
 from semantic_digital_twin.world_description.shape_collection import ShapeCollection
 from semantic_digital_twin.world_description.geometry import Color, Scale, Box
-from semantic_digital_twin.world_description.graph_of_convex_sets import (
+from semantic_digital_twin.world_description.graph_of_convex_sets.boxes import (
     navigation_map_at_target,
 )
 from semantic_digital_twin.world import World
